requirements.txt
customtkinter==5.1.0
pyttsx3==2.7
pyaudio>=0.2.13
numpy==1.26.4
openai==1.43.0
vosk==0.3.30
pygame>=2.1.0
gradio-client==1.3.0
asyncio==3.4.3
AppOpener==1.7
soundfile==0.12.1
openai-whisper==20231117
#torch with cuda for whisper to run fast. In total the venv could take up 6gb. Try here if this version doesn't work. https://pytorch.org/get-started/previous-versions/
torch==2.3.1 -f https://download.pytorch.org/whl/cu121/torch_stable.html
torchvision==0.18.1 -f https://download.pytorch.org/whl/cu121/torch_stable.html
torchaudio==2.3.1 -f https://download.pytorch.org/whl/cu121/torch_stable.html
-f https://download.pytorch.org/whl/cu121/torch_stable.html